renamed random variables
