+++ benchmarks/ivpsolvers.py
@@ -2,7 +2,7 @@
 import numpy as np
 
 from probnum.diffeq import lotkavolterra, probsolve_ivp
-from probnum.random_variables import Constant
+from probnum.randvars import Constant
 
 
 def load_lotkavolterra():
