+++ src/probnum/filtsmooth/gaussfiltsmooth/extendedkalman.py
@@ -6,7 +6,7 @@ import typing
 
 import numpy as np
 
-import probnum.random_variables as pnrv
+import probnum.randvars as pnrv
 import probnum.statespace as pnss
