+++ src/probnum/filtsmooth/gaussfiltsmooth/unscentedkalman.py
@@ -9,7 +9,7 @@ import typing
 
 import numpy as np
 
-import probnum.random_variables as pnrv
+import probnum.randvars as pnrv
 import probnum.statespace as pnss
 import probnum.type as pntype
