+++ src/probnum/problems/_problems.py
@@ -7,7 +7,7 @@ import numpy as np
 import scipy.sparse
 
 import probnum.linops as pnlo
-import probnum.random_variables as pnrv
+import probnum.randvars as pnrv
 import probnum.statespace as pnss
 import probnum.type as pntp
