+++ tests/test_diffeq/test_odefiltsmooth/test_odefiltsmooth.py
@@ -3,7 +3,7 @@ import pytest
 
 from probnum.diffeq import ode
 from probnum.diffeq.odefiltsmooth import KalmanODESolution, probsolve_ivp
-from probnum.random_variables import Constant
+from probnum.randvars import Constant
 
 
 @pytest.fixture
