+++ tests/test_filtsmooth/test_gaussfiltsmooth/test_kalmanposterior.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 
 import probnum.filtsmooth as pnfs
-import probnum.random_variables as pnrv
+import probnum.randvars as pnrv
 import probnum.statespace as pnss
 from probnum._randomvariablelist import _RandomVariableList
