+++ tests/test_randomvariablelist.py
@@ -3,7 +3,7 @@ import unittest
 import numpy as np
 
 from probnum._randomvariablelist import _RandomVariableList
-from probnum.random_variables import Constant
+from probnum.randvars import Constant
 
 
 class TestRandomVariableList(unittest.TestCase):
